Simplifeid promode-temp-vision-state
